backend/requirements.txt
fastapi
uvicorn
sentry-sdk
telnyx
pydantic
python-dotenv
requests